Consolidate CoilModule
