+++ samples/star/src/commonJvm/kotlin/com/slack/circuit/star/di/ApplicationContext.kt
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.slack.circuit.star.di
 
-import android.content.Context
 import javax.inject.Qualifier
 
-/** Qualifier to denote a [Context] that is specifically an Application context. */
+/** Qualifier to denote a `Context` that is specifically an Application context. */
 @Qualifier annotation class ApplicationContext
