+++ samples/star/src/jvmMain/kotlin/com/slack/circuit/star/di/BaseUiModule.kt
@@ -2,22 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.slack.circuit.star.di
 
-import coil3.ImageLoader
 import coil3.PlatformContext
-import coil3.network.NetworkFetcher
 import com.squareup.anvil.annotations.ContributesTo
 import com.squareup.anvil.annotations.optional.SingleIn
 import dagger.Module
 import dagger.Provides
-import io.ktor.client.HttpClient
 
 @ContributesTo(AppScope::class)
 @Module
 object BaseUiModule {
+  @ApplicationContext
   @SingleIn(AppScope::class)
   @Provides
-  fun provideImageLoader(httpClient: dagger.Lazy<HttpClient>): ImageLoader =
-    ImageLoader.Builder(PlatformContext.INSTANCE)
-      .components { add(NetworkFetcher.Factory(lazy { httpClient.get() })) }
-      .build()
+  fun provideAppContext(): PlatformContext = PlatformContext.INSTANCE
 }
